Create Task domain model
Used the rather recently introduced "sealed" keyword to implement this type as a "sum type" to model it as faithfully as possible: I.e., don't just use a regular POJO (which would be a "product type") and thereby also having the field "completionTime" on open tasks, even though it would always be null / empty there.

In the domain, types should be as refined as possible, thus we also use UUID instead of just String for the id field.

Also created a testdata factory for this model to facilitate easy testdata creation in the future.