Create business logic for creating, toggling and retrieving tasks
Also created a repository interface that contains all the methods our business logic requires from some storage mechanism. Note that this interface is part of our business logic and thus belongs to the same domain service layer. Its implementations however will be adapters in the future.

Also added a fake implementation of this interface for now, such that the application context still loads. This will be removed again as soon as we actually implement an adapter related to storage.