Add an optional dueDate property to tasks
This commit shows, how the whole application is affected by such a change.

Note, that in the domain layer we actually use Optional as the field type, even though the JDK developers recommend to only use it as a method return type. However, thanks to our anti-corruption layer we can be sure that the value will never be "null" in the domain layer and thus we actually benefit from using Optional (instead of just adding a third possibility: null, empty, present).