Create a scheduled job that notifies about due tasks
The scheduler and its sender interface are placed in the application layer as typical examples of application internals, that however are not really business logic and thus do not belong to the domain layer.

The implementation of this new interface is created as a new adapter. Note that just logging to the console is of course no "real" notification, however it suffices to demonstrate the architecture. In your production application you might crate an adapter that sends emails, Slack messages or something along those lines.