Configure MongoDB connection an add a 'database mode'
in which the application uses a real MongoDB instance instead of the in-memory cache.

It is configured by changing the application property 'storage.type' from 'cache' to database.

Note, that the application won't start in the new mode yet, as we currently do not have an implementation of TaskRepository for MongoDB.