Provide an implementation of TaskRepository inside the MongoDB adapter
and thereby complete the adapter implementation.

Also, created a JUnit extension that allows writing integration tests against an actual MongoDB instance created via testcontainers.

Reused existing integration and end-to-end tests for the cache mode in the database setting. This avoids a lot of test code duplication and affords reasonable certainty that the two adapters behave the same.